% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoremodel.R
\name{collElimination}
\alias{collElimination}
\title{Collinearity Elimination Automatically}
\usage{
collElimination(df, cor_critical = 0.8)
}
\arguments{
\item{df}{A dataframe only with Xs and Y variable, all the X variables must have been binned, such as after woe-encoded.}

\item{cor_critical}{Specifies the correlation coefficient threshold for extracting, default \code{0.8}.}
}
\value{
A dataframe after elimination
}
\description{
\code{collElimination} is a recursive function, which will recursivly eliminate the collinearity X variables with smaller IV automatically, then return a new dataframe.
}
\details{
for collinearity elimination, recursion is necessary to avoid mistakenly-deleted.
}
\seealso{
Other collinearity elimination functions: \code{\link{collElimination2}}
}
\concept{collinearity elimination functions}
