% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoremodel.R
\name{collElimination2}
\alias{collElimination2}
\title{Collinearity Elimination by Giving Variables' IV}
\usage{
collElimination2(df, IVdf, cor_critical = 0.8)
}
\arguments{
\item{df}{A dataframe only with Xs and Y variable, all the X variables must have been binned, such as after woe-encoded.}

\item{IVdf}{A dataframe of every X variable's IV.}

\item{cor_critical}{Specifies the correlation coefficient threshold for extracting, default \code{0.8}.}
}
\value{
A dataframe after elimination
}
\description{
\code{collElimination2} is a recursive function, which will recursivly eliminate the collinearity X variables with smaller IV. Different from \code{\link{collElimination}}, it must provide parameter \code{IVdf}, so need not recalculate every variable's IV for speed up.
}
\details{
for collinearity elimination, recursion is necessary to avoid mistakenly-deleted.
}
\seealso{
Other collinearity elimination functions: \code{\link{collElimination}}
}
