% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoremodel.R
\name{convertCutPoints}
\alias{convertCutPoints}
\title{Auxiliary Function: Convert Cut Points to Cutbound}
\usage{
convertCutPoints(cutpoints)
}
\arguments{
\item{cutpoints}{A vector with serial cut points, such as \code{c(10,20,30,Inf,NA)}, in this example, Inf and NA is necessary.}
}
\value{
A list length as cutpoints vector
}
\description{
Auxiliary function: convert vector of cut points to a list of cutbound with lbound and ubound.
}
\examples{
convertCutPoints(c(10,20,30,Inf,NA))
}
\seealso{
Other dataset binning and woe-encoding functions: \code{\link{dfBinningFun}},
  \code{\link{executeBinFun_df}},
  \code{\link{genConfigList}}, \code{\link{preBinningFun}},
  \code{\link{smbinning2}}, \code{\link{woeEncodeFun_df}}
}
