% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoremodel.R
\name{convertToFactor}
\alias{convertToFactor}
\title{Convert Character Variables to Factors}
\usage{
convertToFactor(df)
}
\arguments{
\item{df}{A dataframe to be converted.}
}
\value{
A new dataframe with the same order of variables
}
\description{
\code{convertToFactor} will convert all character variables in dataset to factors.
}
\examples{
data(CreditData)
mysample <- convertToFactor(CreditData)
}
