% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoremodel.R
\name{convertType}
\alias{convertType}
\title{Convert Variables to Other Types}
\usage{
convertType(df, toType, vars = -1)
}
\arguments{
\item{df}{A dataframe to be converted.}

\item{toType}{The type converted to be, values must be one of \code{c("fac","cha","int")}. If \code{toType="int"}, the converted result is intercepting the integer part of specified variable, not rounding.}

\item{vars}{Vector of column names or numbers to convert, \code{-1} means to convert all matched variables based on \code{toType} automatically, see details.}
}
\value{
A new dataframe with the same order of variables
}
\description{
\code{convertType} will convert specified variables in dataset to other types.
}
\details{
when \code{vars=-1}, which variables to be converted depends on \code{toType}:
if \code{toType="fac"}, all the character variables will be converted;
if \code{toType="cha"}, all the factor variables will be converted;
if \code{toType="int"}, all the numeric variables will be converted.
}
\examples{
data(CreditData)
sample_fac <- convertType(CreditData, toType="fac")
}
