% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoremodel.R
\name{crossValidation}
\alias{crossValidation}
\title{Cross Validation for Score Model}
\usage{
crossValidation(df, t = 1, k = 10, ifprint = TRUE)
}
\arguments{
\item{df}{A dataframe only with Xs and Y variables, and the last variable must be Y, and all the X variables have been binned, but have not been woe-encoded.}

\item{t}{An integer, means t times, default \code{1}.}

\item{k}{An integer, means k folds, default \code{10}.}

\item{ifprint}{Logical, whether to print the intermediate results, default \code{TRUE}.}
}
\value{
A vector contains average of train and test ks value
}
\description{
\code{crossValidation} is a easy to use function of t times k folds cross validation for score model. The woe will be recomputed at every time and fold validation.
}
\details{
this function will cost much time if you set \code{t} and \code{k} too large, \code{t=1} and \code{k=10} is recommended.
}
\seealso{
Other model performance functions: \code{\link{Curve_Data}},
  \code{\link{myCurves}}, \code{\link{myks}}
}
