% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoremodel.R
\name{Curve_Data}
\alias{Curve_Data}
\title{Auxiliary Function: Create the Appropriate Data Structure for Model Performance Visualization}
\usage{
Curve_Data(y, predict_y, gap = 0.05)
}
\arguments{
\item{y}{A vector of target, only containing 1 and 0.}

\item{predict_y}{A vector of p prediction.}

\item{gap}{A numeric, the gap of two quantile points, default \code{0.05}.}
}
\value{
A list
}
\description{
Auxiliary function: create the appropriate data structure for function \code{\link{myCurves}}, return a list.
}
\details{
the returned list contains two components: a \code{df} 'data.frame(Quant, PropBad, AccumBad, PropGood, AccumGood, BadRate)' and a numeric \code{ks} value.
}
\seealso{
Other model performance functions: \code{\link{crossValidation}},
  \code{\link{myCurves}}, \code{\link{myks}}
}
\concept{model performance functions}
