% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoremodel.R
\name{delFewValues}
\alias{delFewValues}
\title{Delete Variables Based on Number of Different Values}
\usage{
delFewValues(df, minN = 5, minF = 2, exclude = NULL)
}
\arguments{
\item{df}{A dataframe.}

\item{minN}{Min number for numeric variable, default \code{5}.}

\item{minF}{Min number for factor variable, default \code{2}.}

\item{exclude}{Vector of column names or numbers to exclude, default \code{NULL}.}
}
\value{
A dataframe after deleted
}
\description{
\code{delFewValues} will delete variables whose number of different values is too few(see details) to use binning function, and return a new dataframe.
}
\details{
for numeric variable, number of different values must be 5 at least, excluding missing values;
for factor variable, must be 2 at least, excluding missing values also.
}
\examples{
data(CreditData)
mysample <- delFewValues(CreditData, minN = 5, minF = 2, exclude = "target")
}
\seealso{
Other data preprocessing functions: \code{\link{delNArate}},
  \code{\link{delSinvalPercent}}, \code{\link{excludeCol}},
  \code{\link{maxSinvalPercent}}
}
\concept{data preprocessing functions}
