% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoremodel.R
\name{delNArate}
\alias{delNArate}
\title{Compute and Delete NA Rate of Variable}
\usage{
delNArate(df, narate_critical = 0.9, exclude = NULL)
}
\arguments{
\item{df}{A dataframe.}

\item{narate_critical}{A numeric, Specifies NA rate threshold for selecting and returning variables, default \code{0.9}.}

\item{exclude}{Vector of column names or numbers to exclude, default \code{NULL}.}
}
\value{
A list
}
\description{
\code{delNArate} will compute NA rate of every variable in \code{df} firstly, then delete variables whose na rate is greater than or equal to threshold, return a list contains deleted dataframe.
}
\details{
the returned list contains three components, one is dataframe named 'naratedf', second is a dataframe named 'delVardf', last one is an after-deleted dataframe.
}
\examples{
data(CreditData)
result1 <- delNArate(CreditData, narate_critical = 0.7)
result2 <- delNArate(CreditData, narate_critical = 0.7, exclude = "multiloantimes")
mysample <- result2[[3]]
}
\seealso{
Other data preprocessing functions: \code{\link{delFewValues}},
  \code{\link{delSinvalPercent}}, \code{\link{excludeCol}},
  \code{\link{maxSinvalPercent}}
}
