% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoremodel.R
\name{dfBinningFun}
\alias{dfBinningFun}
\title{Generate WOE Configuration List Automatically}
\usage{
dfBinningFun(mydf, binMethod = 1, p = 0.05, maxcat = 10, aliquots = 5)
}
\arguments{
\item{mydf}{A dataframe only with Xs and Y variables, the last column must be Y. All character x variables must be converted to factors in advance.}

\item{binMethod}{An integer from 1 to 4, indicates 4 different binning methods(see details \code{\link{smbinning2}}), default \code{1}.}

\item{p}{A numeric, means percentage of records per bin, from 0 to 0.5, default \code{0.05}.}

\item{maxcat}{An integer, specifies the maximum number of categories, default \code{10}.}

\item{aliquots}{An integer, specifies the number of bins for equal-frequency or equal-interval binning method, default \code{5}.}
}
\value{
A list corresponding to X variables in \code{mydf}
}
\description{
\code{dfBinningFun} is a wrapper and parallelization of \code{\link{smbinning2}}, so it can generate the woe configuration list by using optimal binning automatically.
}
\details{
woe configuration list is a list object lengths as number of x variables in \code{mydf}, whose component is a dataframe consisting of two columns, named as 'vals' and 'woe'.
}
\seealso{
Other dataset binning and woe-encoding functions: \code{\link{convertCutPoints}},
  \code{\link{executeBinFun_df}},
  \code{\link{genConfigList}}, \code{\link{preBinningFun}},
  \code{\link{smbinning2}}, \code{\link{woeEncodeFun_df}}
}
