% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoremodel.R
\name{dfIV}
\alias{dfIV}
\title{Compute Sum IV of All X Variables}
\usage{
dfIV(df)
}
\arguments{
\item{df}{A dataframe only with Xs and Y variable, all the X variables must have been binned in advance, such as after woe-encoded.}
}
\value{
A dataframe contains two columns, one is \code{varname}, another is \code{IV}
}
\description{
\code{dfIV} is a wrapper and parallelization of function \code{\link{sumIV}}, so it can compute the sum IV of every X variable in \code{df}.
}
\details{
notice: all the X variables must have been binned in advance, such as after woe-encoded.
}
\seealso{
Other WOE and IV computing functions: \code{\link{getIV}},
  \code{\link{getWOE}}, \code{\link{sumIV}}
}
