% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoremodel.R
\name{excludeCol}
\alias{excludeCol}
\title{Auxiliary Function: Exclude Columns in Data Frame}
\usage{
excludeCol(df, exclude)
}
\arguments{
\item{df}{A dataframe.}

\item{exclude}{Vector of column names or numbers to exclude.}
}
\value{
A dataframe
}
\description{
Auxiliary function: \code{excludeCol} will exclude the given columns, then return a new dataframe.
}
\examples{
data(CreditData)
x_sample <- excludeCol(CreditData, exclude = "target")
}
\seealso{
Other data preprocessing functions: \code{\link{delFewValues}},
  \code{\link{delNArate}}, \code{\link{delSinvalPercent}},
  \code{\link{maxSinvalPercent}}
}
