% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoremodel.R
\name{executeBinFun}
\alias{executeBinFun}
\title{Auxiliary Function: Execute Binning on An Observation}
\usage{
executeBinFun(x, config_list)
}
\arguments{
\item{x}{A dataframe with only one row, which is the observation to be binned.}

\item{config_list}{A list of woe configuration, whose every component is a dataframe corresponding to a x variable.}
}
\value{
A character vector
}
\description{
Auxiliary function: similar to function \code{\link{woeEncodeFun}}, base on \code{config_list} to execute binning on an observation, but will not proceed woe-encode.
}
\details{
parameter \code{config_list} can be generated by function \code{\link{dfBinningFun}}.
}
\seealso{
Other dataset binning and woe-encoding functions: \code{\link{convertCutPoints}},
  \code{\link{dfBinningFun}},
  \code{\link{executeBinFun_df}},
  \code{\link{genConfigList}}, \code{\link{preBinningFun}},
  \code{\link{smbinning2}}, \code{\link{woeEncodeFun_df}},
  \code{\link{woeEncodeFun}}
}
