% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoremodel.R
\name{executeBinFun_df}
\alias{executeBinFun_df}
\title{Execute Binning on Dataset}
\usage{
executeBinFun_df(x_sample, config_list)
}
\arguments{
\item{x_sample}{A dataframe with only X variables and contains all the observations to be encoded.}

\item{config_list}{A list of woe configuration, whose every component is a dataframe corresponding to a x variable.}
}
\value{
A dataframe after binned
}
\description{
\code{executeBinFun_df} is similar to \code{\link{woeEncodeFun_df}}, based on \code{config_list} to execute binning on all observations, but will not proceed woe-encode.
}
\details{
parameter \code{config_list} can be generated by function \code{\link{dfBinningFun}}.
}
\seealso{
Other dataset binning and woe-encoding functions: \code{\link{convertCutPoints}},
  \code{\link{dfBinningFun}}, \code{\link{genConfigList}},
  \code{\link{preBinningFun}}, \code{\link{smbinning2}},
  \code{\link{woeEncodeFun_df}}
}
\concept{dataset binning and woe-encoding functions}
