% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoremodel.R
\name{extfromFit}
\alias{extfromFit}
\title{Auxiliary Function: Extract Elements From Fit Object}
\usage{
extfromFit(fit, element)
}
\arguments{
\item{fit}{A fitted object of class inheriting from \code{\link{LRfit}} or \code{\link{glm}}.}

\item{element}{Name character of element to be extracted, can be one of c("term","estimate","std.error","statistic","p.value").}
}
\value{
A vector
}
\description{
Auxiliary function: \code{extfromFit} will extract elements from a \code{fit} object, then return a vector.
}
\details{
parameter \code{fit} can be generated by function \code{\link{LRfit}} or \code{\link{glm}}.
if element 'term' is extracted, the extracting result already contains '(Intercept)' at the first one, please remember to remove it if you just need variable names.
}
