% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoremodel.R
\name{fread_basedict}
\alias{fread_basedict}
\title{Read Dataset Based on Data-dictionary}
\usage{
fread_basedict(myfile, mydict, na = c("", ".", "NA", "N/A", "NULL"),
  excludevar = NULL)
}
\arguments{
\item{myfile}{File name character of data set, must be a csv file.}

\item{mydict}{File name character of variable dictionary, must be a csv file.}

\item{na}{A character vector which to be interpreted as NA values, default \code{c("",".","NA","N/A","NULL")}.}

\item{excludevar}{A vector consisting of the variable names want to be excluded in 'mydict' file, default \code{NULL}.}
}
\value{
A dataframe
}
\description{
\code{fread_basedict} is a wrapper of \code{fread}, which allows you read dataset based on your data-dictionary(see details).
}
\details{
the format of data-dictionary is required:
1.the first column must be names of variables;
2.the second column must be labels of variables;
3.the third column must be flags which indicate whether it is a variable;
4.the fourth column must be types of variables, and the value in this column should be 'VARCHAR', 'NUMBER' or NULL.
}
