% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoremodel.R
\name{genConfigList}
\alias{genConfigList}
\title{Generate WOE Configuration List Manually}
\usage{
genConfigList(df, cut_list)
}
\arguments{
\item{df}{A dataframe only with Xs and Y variables, and the last variable must be Y. Y must be 0 or 1, with 0 represents good and 1 represents bad.}

\item{cut_list}{A list of cut points vector corresponding X variables, but for factor variable, the element is fixed to constant 'factor'.}
}
\value{
A list corresponding to X variables in \code{df}
}
\description{
\code{genConfigList} can be used to generate the woe configuration list by batch computing WOEs with given binnings.
}
\details{
\code{cut_list} must contains max and Inf value in every continuous-representing component. Component such as \code{c(10,20,30,Inf,NA)}.
}
\examples{
data(CreditData)
mysample <- CreditData[c("gender","gscore","target")]
genConfigList(mysample, cut_list = list("factor",c(534,566,577,617,Inf,NA)))
}
\seealso{
Other dataset binning and woe-encoding functions: \code{\link{convertCutPoints}},
  \code{\link{dfBinningFun}},
  \code{\link{executeBinFun_df}},
  \code{\link{preBinningFun}}, \code{\link{smbinning2}},
  \code{\link{woeEncodeFun_df}}
}
\concept{dataset binning and woe-encoding functions}
