% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoremodel.R
\name{getEqualFreqCuts}
\alias{getEqualFreqCuts}
\title{Auxiliary Function: Improved Equal-Frequency Binning}
\usage{
getEqualFreqCuts(x, n)
}
\arguments{
\item{x}{A numeric vector you want to bin.}

\item{n}{The number of bins for Equal-Frequency binning.}
}
\value{
A vector
}
\description{
Auxiliary function: \code{getEqualFreqCuts} will get the improved Equal-Frequency binning cut points, return a vector.
}
\examples{
x <- c(1,2,3,3,3,3,3,4,6)
as.vector(quantile(x, probs=seq(0,1,1/3)))
getEqualFreqCuts(x,3)
}
