% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoremodel.R
\name{getIV}
\alias{getIV}
\title{Compute the Given Binning IV}
\usage{
getIV(df, cutbound, totalgood = NULL, totalbad = NULL)
}
\arguments{
\item{df}{A dataframe only with two variables, X and Y.}

\item{cutbound}{A single value for factor variable, or a vector with lbound and ubound for numeric variable.}

\item{totalgood}{Default \code{NULL}. if not default, will use them directly, otherwise will recalculate based on \code{df}.}

\item{totalbad}{Default \code{NULL}. if not default, will use them directly, otherwise will recalculate based on \code{df}.}
}
\value{
A numeric
}
\description{
\code{getIV} will compute the given binning IV of x variable, return a numeric.
}
\examples{
data(CreditData)
getIV(CreditData[c("gender","target")], cutbound = "M")
getIV(CreditData[c("gscore","target")], cutbound = c(-Inf,534))
getIV(CreditData[c("gscore","target")], cutbound = c(577,617))
getIV(CreditData[c("gscore","target")], cutbound = c(617,Inf))
getIV(CreditData[c("gscore","target")], cutbound = NA)
}
\seealso{
Other WOE and IV computing functions: \code{\link{dfIV}},
  \code{\link{getWOE}}, \code{\link{sumIV}}
}
