% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoremodel.R
\name{getWOE}
\alias{getWOE}
\title{Compute the Given Binning WOE}
\usage{
getWOE(df, cutbound, totalgood = NULL, totalbad = NULL)
}
\arguments{
\item{df}{A dataframe only with two variables, X and Y.}

\item{cutbound}{A single value for factor variable, or a vector with lbound and ubound for numeric variable.}

\item{totalgood}{Default \code{NULL}. if not default, will use them directly, otherwise will recalculate based on \code{df}.}

\item{totalbad}{Default \code{NULL}. if not default, will use them directly, otherwise will recalculate based on \code{df}.}
}
\value{
a numeric
}
\description{
\code{getWOE} will compute the given binning WOE of x variable, return a numeric.
}
\examples{
data(CreditData)
getWOE(CreditData[c("gender","target")], cutbound = "M")
getWOE(CreditData[c("gscore","target")], cutbound = c(-Inf,534))
getWOE(CreditData[c("gscore","target")], cutbound = c(577,617))
getWOE(CreditData[c("gscore","target")], cutbound = c(617,Inf))
getWOE(CreditData[c("gscore","target")], cutbound = NA)
}
\seealso{
Other WOE and IV computing functions: \code{\link{dfIV}},
  \code{\link{getIV}}, \code{\link{sumIV}}
}
\concept{WOE and IV computing functions}
