% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoremodel.R
\name{insertElement}
\alias{insertElement}
\title{Auxiliary Function: Insert Elements into Vector}
\usage{
insertElement(vec, indexs, values)
}
\arguments{
\item{vec}{A vector to be inserted into.}

\item{indexs}{A vector consisting of location indexs corresponding to the location of new elements after inserted.}

\item{values}{A vector to be inserted into parameter \code{vec}.}
}
\value{
A vector
}
\description{
Auxiliary function: \code{insertElement} can insert specified elements into the specified location of vector, then return a new vector.
}
\details{
the elements in \code{indexs} must be sorted by ascending.
}
\examples{
insertElement(c(6,8,10),c(2,4),c(7,9))
}
