% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoremodel.R
\name{LRfit}
\alias{LRfit}
\title{Logistic Regression Model Training by Stepwise}
\usage{
LRfit(df, sig = 0.05)
}
\arguments{
\item{df}{A dataframe only with Xs and Y variable, all the category X variables must have been encoded(such as WOE, One-Hot and Dumb-Variable encoding).}

\item{sig}{A numeric of significance-level.}
}
\value{
Same as \code{\link{glm}}
}
\description{
\code{LRfit} is a simple wrapper of \code{\link{glm}} for logistic regression, you only need to specify the data and significance-level for training.
}
