% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoremodel.R
\name{LRpredict}
\alias{LRpredict}
\title{Logistic Regression Model Predicting}
\usage{
LRpredict(fit, newdata = NULL)
}
\arguments{
\item{fit}{A fitted object of class inheriting from \code{\link{LRfit}} or \code{\link{glm}}.}

\item{newdata}{A new dataset to be predicted, which only contain x variables, and must have been woe-encoded. Default \code{NULL}, means predicting on train.}
}
\value{
A numeric vector
}
\description{
\code{LRpredict} is a wrapper of \code{\link{predict}}, which can predict probabilities on train or test set. Different from \code{\link{rawPredictFun_df}}, it does not need to give model coefficient and woe configuration list, but the given \code{newdata} must be woe-encoded in advance when predicts on test.
}
\details{
parameter \code{fit} can be generated by function \code{\link{LRfit}} or \code{\link{glm}}.
}
\seealso{
Other model prediction functions: \code{\link{rawPredictFun_df}},
  \code{\link{rawPredictFun}}, \code{\link{transformScore}}
}
