% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoremodel.R
\name{maxSinvalPercent}
\alias{maxSinvalPercent}
\title{Get the Max Percents of All Variable's Single-Value}
\usage{
maxSinvalPercent(df)
}
\arguments{
\item{df}{A dataframe.}
}
\value{
A vector
}
\description{
\code{maxSinvalPercent} is a wrapper of auxiliary function \code{maxSinvalPercent_x}, so it will return the max percent vector of every variable's Single-Value in \code{df} at one time.
}
\seealso{
Other data preprocessing functions: \code{\link{delFewValues}},
  \code{\link{delNArate}}, \code{\link{delSinvalPercent}},
  \code{\link{excludeCol}},
  \code{\link{maxSinvalPercent_x}}
}
