% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoremodel.R
\name{maxSinvalPercent_x}
\alias{maxSinvalPercent_x}
\title{Auxiliary Function: Get the Max Percent of the Given Variable's Single-Value}
\usage{
maxSinvalPercent_x(df, x_nm)
}
\arguments{
\item{df}{A dataframe.}

\item{x_nm}{Name of x variable.}
}
\value{
A numeric
}
\description{
Auxiliary function: return the max percent of Single-Value in the given x variable.
}
\examples{
data(CreditData)
maxSinvalPercent_x(CreditData, "gender")
}
\seealso{
Other data preprocessing functions: \code{\link{delFewValues}},
  \code{\link{delNArate}}, \code{\link{delSinvalPercent}},
  \code{\link{excludeCol}}, \code{\link{maxSinvalPercent}}
}
