% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoremodel.R
\name{myCurves}
\alias{myCurves}
\title{Model Performance Visualization}
\usage{
myCurves(ytrain, predict_ytrain, ytest, predict_ytest, ontest = TRUE,
  lift_bins = 10, P0 = 600, PDO = 50, color_scheme = 1, ifsave = TRUE)
}
\arguments{
\item{ytrain}{A vector corresponding the response variable encoded with 0 and 1 of train set.}

\item{predict_ytrain}{A vector of predicted probability values of train set.}

\item{ytest}{A vector corresponding the response variable encoded with 0 and 1 of test set.}

\item{predict_ytest}{A vector of predicted probability values of test set.}

\item{ontest}{Logical, decide plot on train or on test, default \code{TRUE}.}

\item{lift_bins}{An integer, set the number of binnings, default \code{10}.}

\item{P0}{An integer, the base score, default \code{600}.}

\item{PDO}{An integer, the scale, default \code{50}.}

\item{color_scheme}{An integer, choice the color scheme 1, 2 or 3, default scheme \code{1}.}

\item{ifsave}{Logical, whether to save the chart as a file, default \code{TRUE}.}
}
\description{
\code{myCurves} is an integrated, flexible as well as easy to use function of visualization for model performance.
}
\details{
the result graph has four subgraphs which constitute a matrix, top left is ROC Curve, top right is Lift Figure, lower left is K-S Curve, and lower right is Distribution Figure of standard score.
}
\seealso{
Other model performance functions: \code{\link{Curve_Data}},
  \code{\link{crossValidation}}, \code{\link{myks}}
}
