% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoremodel.R
\name{myks}
\alias{myks}
\title{Compute KS Value of Score Model}
\usage{
myks(y, predict_y)
}
\arguments{
\item{y}{A vector of target, only containing 1 and 0.}

\item{predict_y}{A vector of p prediction.}
}
\value{
A numeric
}
\description{
\code{myks} will compute KS value by giving target and prediction, then return a numeric.
}
\seealso{
Other model performance functions: \code{\link{Curve_Data}},
  \code{\link{crossValidation}}, \code{\link{myCurves}}
}
