% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoremodel.R
\name{preBinningFun}
\alias{preBinningFun}
\title{Variable Pre-Binning, then Computing WOE and IV}
\usage{
preBinningFun(mydata, binMethod = 1, p = 0.05, aliquots = 5,
  mydict = NULL)
}
\arguments{
\item{mydata}{A data frame of dataset consisting of only Xs and Y variables, the last column must be Y. All character x variables must be converted to factors in advance.}

\item{binMethod}{An integer from 1 to 5, indicates 5 different binning methods(see details), default \code{1}.}

\item{p}{A numeric, means percentage of records per bin, from 0 to 0.5, default \code{0.05}.}

\item{aliquots}{An integer, specifies the number of bins for equal-frequency or equal-interval binning method, default \code{5}.}

\item{mydict}{Optional, default \code{NULL}. File name character of variable dictionary with csv file extension, or a dataframe representing the variable dictionary, see details of this parameter in \code{\link{fread_basedict}}.}
}
\value{
A list
}
\description{
\code{preBinningFun} is often used to pre-binning with returning plots and a lots of details for variables filtering, and it has integrated 5 different binning methods(see details). Important notes: the calculated WOE value in this function is the opposite of the actual value.
}
\details{
\code{binMethod=c(1,2,3,4,5)}, means:
  1 means optimal binning, and equal-frequency binning is an alternative when optimal binning is not available.
  2 means optimal binning, and equal-interval binning is an alternative when optimal binning is not available.
  3 means equal-frequency binning.
  4 means equal-interval binning.
  5 means optimal binning only.

this function will generate four files in current directory, including 'binGraph.pdf', 'varSummary.csv', 'summaryIV.csv' and 'insignificantVars.csv'(if it exists), as well as mass csv files in '~/binDetails/' subdirectory. If the subdirectory does not exist, it will be created automatically.
}
\seealso{
Other dataset binning and woe-encoding functions: \code{\link{convertCutPoints}},
  \code{\link{dfBinningFun}},
  \code{\link{executeBinFun_df}},
  \code{\link{genConfigList}}, \code{\link{smbinning2}},
  \code{\link{woeEncodeFun_df}}
}
