% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoremodel.R
\name{rawPredictFun}
\alias{rawPredictFun}
\title{Auxiliary Function: Predict p Value of A Raw Observation}
\usage{
rawPredictFun(x, config_list, coef_vec)
}
\arguments{
\item{x}{A data frame with only one row, which is the observation to be predicted, and it only contain X variables.}

\item{config_list}{A list of woe configuration, whose every component is a dataframe corresponding to a X variable.}

\item{coef_vec}{A vector of model coefficients.}
}
\value{
A numeric
}
\description{
Auxiliary function: \code{rawPredictFun} will predict the p value of an observation, based on woe configuration list(see details) and model coefficients. Different from \code{\link{LRpredict}}, this function can predict on one raw observation directly which have not been woe-encoded or binned.
}
\details{
parameter \code{config_list} can be generated by function \code{\link{dfBinningFun}}.
}
\seealso{
Other model prediction functions: \code{\link{LRpredict}},
  \code{\link{rawPredictFun_df}},
  \code{\link{transformScore}}
}
