% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoremodel.R
\name{splitData}
\alias{splitData}
\title{Splitting Dataset into Train and Test Set}
\usage{
splitData(df, size, ifpercent = TRUE)
}
\arguments{
\item{df}{A dataframe to be splitted.}

\item{size}{A numeric. Specifies the sampling fraction or quantity of train set.}

\item{ifpercent}{Logical, indicating whether the sampling fraction or quantity to be specified.}
}
\value{
A list contains train and test set, and train set is the fisrt.
}
\description{
\code{splitData} will split the dataset into train and test set by given sampling fraction or quantity, then return a list.
}
\examples{
data(CreditData)
splitresult <- splitData(CreditData, size = 0.7, ifpercent = TRUE)
train <- splitresult[[1]]
test <- splitresult[[2]]
}
