% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoremodel.R
\name{sumIV}
\alias{sumIV}
\title{Auxiliary Function: Compute Sum IV of Given Variable}
\usage{
sumIV(df, x, totalgood = NULL, totalbad = NULL)
}
\arguments{
\item{df}{A dataframe only with Xs and Y variable, and the last variable must be Y.}

\item{x}{A name character of X variable to be computed.}

\item{totalgood}{Default \code{NULL}. if not default, will use them directly, otherwise will recalculate based on \code{df}.}

\item{totalbad}{Default \code{NULL}. if not default, will use them directly, otherwise will recalculate based on \code{df}.}
}
\value{
A numeric
}
\description{
Auxiliary function: \code{sumIV} will compute the sum IV of a binned X variable, return a numeric.
}
\details{
notice: the given X variable must have be binned in advance.
}
\examples{
data(CreditData)
sumIV(CreditData, x = "gender")
}
\seealso{
Other WOE and IV computing functions: \code{\link{dfIV}},
  \code{\link{getIV}}, \code{\link{getWOE}}
}
\concept{WOE and IV computing functions}
