% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoremodel.R
\name{transformScore}
\alias{transformScore}
\title{Transform Prediction to Standard Score}
\usage{
transformScore(p, theta0, P0 = 600, PDO = 50)
}
\arguments{
\item{p}{A numeric vector, the predicted probability.}

\item{theta0}{A numeric, the base ratio of bad:good in train.}

\item{P0}{An integer, the base score, default \code{600}.}

\item{PDO}{An integer, the scale, default \code{50}.}
}
\value{
A integer vector
}
\description{
\code{transformScore} will transform the predicted probability to standard score by giving base score and scale, then return a integer vector.
}
\details{
parameter \code{theta0} must be the bad:good of train at any time.
}
\seealso{
Other model prediction functions: \code{\link{LRpredict}},
  \code{\link{rawPredictFun_df}}
}
\concept{model prediction functions}
