% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoremodel.R
\name{woeEncodeFun}
\alias{woeEncodeFun}
\title{Auxiliary Function: Encode An Observation as WOE}
\usage{
woeEncodeFun(x, config_list)
}
\arguments{
\item{x}{A dataframe with only one row, which is the observation to be encoded.}

\item{config_list}{A list of woe configuration, whose every component is a dataframe corresponding to a x variable.}
}
\value{
A numeric vector
}
\description{
Auxiliary function: \code{woeEncodeFun} will encode an observation as woe values, then return a numeric vector.
}
\details{
parameter \code{config_list} can be generated by function \code{\link{dfBinningFun}}.
}
\seealso{
Other dataset binning and woe-encoding functions: \code{\link{convertCutPoints}},
  \code{\link{dfBinningFun}},
  \code{\link{executeBinFun_df}},
  \code{\link{executeBinFun}}, \code{\link{genConfigList}},
  \code{\link{preBinningFun}}, \code{\link{smbinning2}},
  \code{\link{woeEncodeFun_df}}
}
