% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoremodel.R
\name{woeEncodeFun_df}
\alias{woeEncodeFun_df}
\title{Encode the Entire Dataset as WOE}
\usage{
woeEncodeFun_df(x_sample, config_list)
}
\arguments{
\item{x_sample}{A dataframe with only X variables and contains all the observations to be encoded.}

\item{config_list}{A list of woe configuration, whose every component is a dataframe corresponding to a x variable.}
}
\value{
A dataframe after woe-encoded
}
\description{
\code{woeEncodeFun_df} is a wrapper and parallelization of \code{\link{woeEncodeFun}}, so you can use it to encode the entire dataset at one time.
}
\details{
parameter \code{config_list} can be generated by function \code{\link{dfBinningFun}}.
}
\seealso{
Other dataset binning and woe-encoding functions: \code{\link{convertCutPoints}},
  \code{\link{dfBinningFun}},
  \code{\link{executeBinFun_df}},
  \code{\link{executeBinFun}}, \code{\link{genConfigList}},
  \code{\link{preBinningFun}}, \code{\link{smbinning2}},
  \code{\link{woeEncodeFun}}
}
